Post-move import fixes
