+++ app/src/main/AndroidManifest.xml
@@ -44,7 +44,7 @@
             android:screenOrientation="portrait"
             android:theme="@style/ParticleSetupTheme.NoActionBar.TranslucentSystemBars" />
         <activity
-            android:name="io.particle.android.sdk.ui.DeviceListActivity"
+            android:name="io.particle.android.sdk.ui.devicelist.DeviceListActivity"
             android:label="@string/title_device_list"
             android:screenOrientation="portrait"
             android:theme="@style/ParticleSetupTheme.NoActionBar.TranslucentSystemBars">
@@ -62,12 +62,12 @@
         <activity
             android:name="io.particle.android.sdk.ui.InspectorActivity"
             android:label="@string/title_inspector_detail"
-            android:parentActivityName="io.particle.android.sdk.ui.DeviceListActivity"
+            android:parentActivityName="io.particle.android.sdk.ui.devicelist.DeviceListActivity"
             android:screenOrientation="portrait"
             android:theme="@style/TinkerSetup">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="io.particle.android.sdk.ui.DeviceListActivity" />
+                android:value="io.particle.android.sdk.ui.devicelist.DeviceListActivity" />
         </activity>
         <!--
         All of the following are from the device setup lib, and must be present in your app's
