+++ app/src/main/java/io/particle/android/sdk/ui/devicelist/Comparators.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.particle.android.sdk.ui;
+package io.particle.android.sdk.ui.devicelist;
 
 import java.util.ArrayList;
 import java.util.BitSet;
